--- TeamTextRPG/Event.cs
@@ -1,12 +1,0 @@
-﻿using System;
-namespace TeamTextRPG
-{
-	public class Event
-	{
-		public Event()
-		{
-			Console.WriteLine("Don't touch anything!!");
-		}
-	}
-}
-
